Add tests for fetcher, refactor
