+++ package.json
@@ -13,15 +13,17 @@
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.11.6",
     "@types/node-fetch": "^2.5.12",
+    "@types/chai": "^4.2.22",
+    "@types/sinon": "^10.0.6",
     "chai": "^4.3.4",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
     "ts-node": "^10.4.0",
     "ts-standard": "^10.0.0",
-    "typescript": "^4.4.4"
+    "typescript": "^4.4.4",
+    "sinon": "^12.0.1"
   },
   "dependencies": {
-    "@types/chai": "^4.2.22",
     "js-yaml": "^4.1.0",
     "node-fetch": "^2.6.6",
     "uri-js": "^4.4.1"
