+++ src/test/data/hellofield.yml
@@ -1,5 +1,5 @@
 {
 "name": "hello",
-"doc": {"$include": "hello.txt"},
+"doc": {"$include": "test.txt"},
 "type": "string"
 }
